pyproject.toml
[tool.poetry]
name = "modbus-utility"
version = "0.2.0"
description = ""
authors = ["jobenas <jobenas@energyatech.com>"]
readme = "README.md"

[tool.poetry.scripts]
modbus-utility = "modbus_utility.main:app"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.15.1"
pyserial = "^3.5"
pydantic = "^2.9.2"
ruff = "^0.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
